+++ CHANGES.md
@@ -22,6 +22,7 @@
 - HidesVariableCheck now allows for variable definition in ancestor scope, when defined at lower line.
 - Add ForbiddenInheritanceCheck.
 - TrailingWhitespaceCheck now marks the actual whitespace.
+- Add defintion functionality to magik-language-server.
 - Various small fixes.
 
 0.7.1 (2023-02-21)